pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Neel Gokhale <neel@cohere.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cohere = "^5.9.1"
langchain-cohere = "^0.2.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
